// src/enums/role.enum.ts
export enum Role {
  ADMIN = 'Administrator', // Quản trị viên: Quản lý toàn bộ hệ thống và người dùng.
  MANAGER = 'Manager', // Quản lý: Quản lý một hoặc nhiều chức năng như sản phẩm, đơn hàng, hoặc nội dung.
  STAFF = 'Staff', // Nhân viên: Thực hiện các nhiệm vụ hàng ngày, hỗ trợ quản lý và khách hàng.
  CUSTOMER = 'Customer', // Khách hàng: Người dùng có thể mua sản phẩm và sử dụng dịch vụ.
}